Fix automatic hiding of battery monitor
This commit makes Conky display the current energy of BAT1 when the batterys
reported status is "Charging" or "Discharging" and hides it otherwise.

Apparently my battery has a funny habit of going into an "unknown" status
($ cat /sys/class/power_supply/BAT1/status) when it's almost, but not quite
fully charged, and can't be bothered to be charged any more. This happened after
reconnecting it before it discharged much (was at about 95%).

However, once charging gets started it also gets completed as far as I can see.