Set GPG_TTY in `.bashrc`, not `.profile`
This environment variable is used by gpg-agent.  For example, to display
the curses-based `pinentry` program in the terminal from which the agent
was accessed.

Setting it in `.profile`, like I did, means it was only set and exported
by login shells.  That's bad.  Usually it was `/dev/tty1` everywhere,
since that's where I log in on my Arch system.