systemd manages wlan0 now: adjust some scripts
I will use systemd to start `wpa_supplicant` and `dhcpcd` on Arch now
(with `wpa_supplicant@wlan0` and `dhcpcd@wlan0` units).  I used to start
both manually in specific screen windows.

Make some changes to the `wifail` and `startup.sh` scripts to account
for this and add commands to quickly check the output of
`wpa_supplicant` and `dhcpcd` to `snippets.bash`.