Make starting Mutt more convenient
Instead of having Mutt prompt for my password and entering it directly
or by using passmenu [1], make Mutt invoke pass(1) directly.  The result
is a `pinentry-curses` dialog (or Mutt starting directly if the
passphrase is cached).

Until now I would start passmenu manually, select the password, enter
the passphrase for GPG (`pinentry-gtk-2` is used), and hit Enter a
second time to confirm the password.

[1]: https://git.zx2c4.com/password-store/tree/contrib/dmenu/passmenu