Merge X resource files individually from xinitrc
Instead of `#include`ing X resource files below `~/.Xresources.d/` from
`~/.Xresources` [1][2], modify `xinitrc` to merge them individually.
This allows creating symlinks for only a subset of the program-specific/
topical files below `Xresources.d/`.

Other advantages of this approach are that file names don't need to be
repeated in the "master" Xresources file and that no C preprocessor is
required.

[1]: https://wiki.archlinux.org/index.php/x_resources#Include_files
[2]: https://www.reddit.com/r/unixporn/wiki/organizing_xresources