mn: avoid creating links with consecutive slashes
If the second argument is a directory name with a trailing slash (which
may be added by tab completion), remove the slash before appending the
basename of the source file.

Before, when the second argument was a directory name with a trailing
slash, the target path `ln` was invoked with would have two consecutive
slashes (unless rewritten inside the `case` statement).

The consecutive slashes are inconsequential (they are treated as one
slash [1]) but look ugly.

[1]: https://unix.stackexchange.com/a/1919