Tune keybinds to hush "This IS window [n]" message
Modify the xbindkeys(1) keybindings for starting or focusing Mutt,
Irssi, and Neovim to make screen(1) switch to the "blank window" before
switching to the target window.  The point is to work around screen
displaying a message informing us that the requested window was already
active (if that's the case), which stalls any updates of the virtual
terminal while shown.