Firefox: style Hacker News with userContent.css
I've used this CSS with Stylish [1] and later Stylus [2] for a while.  I
don't know where I originally got it, but remember that I tweaked it a
bit.  I think it will be more convenient to track it as part of my
dotfiles and make it work without a browser extension.  At the moment,
not all rules are applied as intended (see [3]).

[1]: https://en.wikipedia.org/wiki/Stylish
[2]: https://github.com/openstyles/stylus
[3]: https://www.reddit.com/comments/7c7r3k//dpnvo2c