XBindKeys: change master instead of PCM volume
I think PulseAudio sets the PCM volume to the maximum when headphones
are plugged in or out, and lowers the master volume so the total volume
is kept at the same level it was at before.  This means increasing the
PCM volume is not possible after plugging headphones in or out: modify
`xbindkeysrc` so the multimedia keys for changing the sound volume use
the master instead of the PCM volume.