Fix snippet for adding all songs to MPD queue
Fix the snippet that is supposed to first clear and then add all songs
to MPD's queue.  Apparently, mpc(1) used to suggest to "[u]se
`mpc ls | mpc add` to add all files to the playlist".  This is no longer
the case.  My local copy of mpc(1) now says to "[u]se `mpc add /` to add
all files to the queue".

The problem with `mpc ls` is that it also lists playlists.