Screenshot: convert `title` to `alt` attribute
I was doing it wrong.  See [1] and [2].

"The `alt` attribute contains a textual description of the image, which
isn't mandatory but is incredibly useful for accessibility".

"An `alt` attribute's value should clearly and concisely describe the
image's content."

"You should [...] provide a useful value for `alt` whenever possible."

"The `title` attribute is not an acceptable substitute for an `alt`
attribute.  Additionally, avoid duplicating the `alt` attribute's value
in a `title` attribute declared on the same image."

[1]: https://daringfireball.net/projects/markdown/syntax#img
[2]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/img