Add commands concerning rsync.net to snippets.bash
Add some commands for doing stuff with my rsync.net cloud storage to
`snippets.bash`.  The new SSH configuration file sets the host and user
name for "esgaroth" and is encrypted with [git-crypt][1].  This SSH
configuration has to be available in `/root/.ssh/` for the "sudo rsync"
command to work.

[1]: https://rsync.net
[2]: https://github.com/AGWA/git-crypt