Update start/focus keybinds for Neovim and Neomutt
As already done in the Irssi keybind, use screen's `msgwait` command
instead of `select -`.  The purpose of both commands is to make the
keybinds behave nicely when the right screen window is already selected
(no "This IS window 0 (nvim)." message).  The `msgwait` method is nicer
because it doesn't cause any flickering; the `select -` method does.

Also make some minor unrelated tweaks to the keybinds.