NeoMutt: start moving toward a multi-account setup
I want to start using [FastMail][1], but I probably won't be able to
stop using Gmail for a long, long time.  We'll have to make some changes
to the NeoMutt configuration to enable using multiple accounts.  See [2]
and [3].

This commit only moves some Gmail-specific settings into a dedicated
file as a first step.

[1]: https://en.wikipedia.org/wiki/FastMail
[2]: https://wiki.archlinux.org/index.php/Mutt#Multiple_accounts
[3]: https://neomutt.org/guide/optionalfeatures.html#account-hook