Bash PS1: fix `\$` not being expanded
`\$` is explained in the PROMPTING section of bash(1).  I used it
wrongly, and when I ran `su` or `sudo su` or `sudo -s`, I still got a
prompt that ended in $ rather than #.

What I did before was broken because the expansion described in the
bash(1) PROMPTING section happens before command substitution.