NeoMutt: add separate "i" keybind for FastMail
Use separate "i" keybinds for Gmail and FastMail.  The old keybind
copies the selected message to the INBOX mailbox.  This worked (and
continues to work) well with Gmail (presumably because Gmail's mailboxes
act like tags), but did silly things with FastMail: copying a message to
the INBOX mailbox with "i" and later archiving the copy with "e"
duplicated the message.