Add XBindKeys keybind for quickly sending an email
I think it will be nice to have a workflow for composing and sending an
email that doesn't involve opening NeoMutt first.  The problem when
opening NeoMutt (or any MUA) first is that lots of messages are shown
before starting to compose a new email is possible.  This poses an
unnecessary distraction.