Stop using vim-plug
Use Git submodules and Vim's built-in "packages" feature (for
'runtimepath' management) instead of vim-plug.

See `:h packages`, `:h :packadd`, and `:h pack-add`.  Also see [1], [2],
[3], and [4].  I like to get crazy.

The startup times of Vim and Neovim are slightly worse now.  Almost all
of the extra time is due to pastery.vim [5] not being loaded lazily
anymore.

The main advantage of Vim's built-in "packages" feature is that it's
built in.  The main advantage of using Git submodules is that we can get
exactly the same version of all plugins everywhere.

[1]: https://shapeshed.com/vim-packages/
[2]: https://vi.stackexchange.com/q/9522
[3]: https://github.com/junegunn/vim-plug
[4]: https://github.com/tpope/vim-pathogen
[5]: https://github.com/skorokithakis/pastery.vim