Add dictd.conf
I had to copy this file to `/etc/dict/dictd.conf`.  Using a symbolic
link doesn't work.

My reason for adding this file to my dotfiles repository is that I
modified the order of the databases.  The order matters because "dictd
returns definitions from dictionary databases in the order in which they
are listed in its configuration file." [1]

I'm not sure whether anything automatically overwrites
`/etc/dict/dictd.conf`.  We'll see.

[1]: https://manpages.debian.org/testing/dictd/dictdconfig.8.en.html