Swap return and backslash keybinds
I have a new ThinkPad X220 and this one has a US keyboard.  My other
X220 has a German keyboard which means the backslash key is where the
return key should be when using the US keyboard layout.

Swapping the return and backslash keybinds means everything stays more
or less like it was on my other ThinkPad X220.