Create shoddy `pinentry-spawn-xterm` script
This script may be shoddy (well, it is), but at least, unlike
`pinentry-gtk-2`, it can deal with high-pixel-density displays.  And,
unlike `pinentry-gnome3`, it works without GNOME.  And, unlike
`pinentry-tty` and `pinentry-curses`, it can be used when the process
that invoked GPG isn't connected to a (big-enough) terminal.

https://wiki.archlinux.org/title/GnuPG#pinentry
https://archlinux.org/packages/core/x86_64/pinentry/
https://gnupg.org/related_software/pinentry/
https://github.com/gpg/pinentry
https://www.linuxfromscratch.org/blfs/view/cvs/general/pinentry.html
https://unix.stackexchange.com/q/618341
https://www.gnupg.org/documentation/manuals/assuan/
https://github.com/t-8ch/pinentry-bemenu

Also see `info pinentry`.