Super+F1: also show man pages from sectns. 6 and 8
There's no rhyme or reason to which man pages are in section 6 or 8
instead of 1.  For example, `figlet` is in section 6, but `cowsay` is in
section 1.  And `ping` is in section 8, whereas `dig` is in section 1.

Having to remember these arbitrary classifications and pointless
distinctions is silly, so let's pretend they don't exist for the
purposes of the Super+F1 keybind.