Add a new command type to fzf-command-runner
The new `less` command type is a minimal variation of the `follow`
command type.  The former differs from the latter only in that the
command isn't echoed.  In the case of `git -C ~/bash-history log -U0`,
the command was shown for a mere instant prior to `less` fully taking
over the terminal, which was unhelpful and looked slightly glitchy.