Super+I: replace `mktemp` with `qualified-noun`
The names that `mktemp` generates are rough.  With `qualified-noun`, we
can get perfectly cromulent ones.

This commit also switches from `/tmp/` to `~/tmp/`.  The former uses
volatile memory (at least on my system), and I don't like the potential
for unplanned disappearance this entails, even for "temporary" files.
Or at least not for those I create manually.  Furthermore, `/tmp/` is
full of footguns [1].

I'm not sure what to do with `instagit`, so it's gone for now.

[1]: https://dotat.at/@/2024-10-22-tmp.html